// frontend/src/app/api/models/create-tour-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateTourRequest {
  description?: (string | null);
  endLat?: (number | string);
  endLng?: (number | string);
  endLocation?: string;
  mapImagePath?: (string | null);
  startLat?: (number | string);
  startLng?: (number | string);
  startLocation?: string;
  title?: string;
  transportType?: string;
}